wasm-jit: emit New/UintMulHigh, fix pointer-width field access, seed EC and __main__
Compiled-trace correctness fixes for the wasm32 full-JIT backend.

majit-backend-wasm:
- Emit New/NewWithVtable/NewArray/NewArrayClear (previously skipped as
  no-ops). Route allocation through the jit_call host trampoline to the
  new wasm_jit_alloc / wasm_jit_alloc_array module helpers; write the
  ob_type field with a 4-byte i32 store at vtable_offset.
- Implement UintMulHigh via a 32-bit split expansion using reserved
  scratch locals; without it, strength-reduced division/modulo by a
  constant produced wrong results.
- Extend has_call_ops so allocation ops pull in the jit_call import.

majit-backend / majit-metainterp:
- Add the residual_host_call hook (call_stub.rs) and route call_int/
  call_void and the bh_call_* defaults through it so residual calls
  reflect their real wasm signature; make the bh setters/getters real.

Pointer-width (wasm32) fixes for fields read/written as raw memory:
- virtualizable.rs: vable token and Ref/Signed field & array-item
  access at usize/isize width.
- pyre-jit-trace frame_layout.rs (vable_token clear) and state.rs
  (pyobject array descr item_size).

pyre-wasm run_python_impl:
- set_last_exec_ctx before eval so a blackhole bh_call_fn_impl resolves
  its parent frame.
- Register the __main__ module in sys.modules.

pyre-wasm-runner: log the entered call area in the trampoline (no-op
unless tracing).

check.py --backend wasm: 90 passed / 23 failed (was JIT-panic on every
hot loop). Native dynasm/cranelift unchanged.

+++ pyre/pyre-jit-trace/src/frame_layout.rs
@@ -108,7 +108,9 @@ unsafe extern "C" fn pyre_clear_vable_token(obj_ptr: i64) {
     unsafe {
         let ptr = obj_ptr as *mut u8;
         if !ptr.is_null() {
-            let token_ptr = ptr.add(PYFRAME_VABLE_TOKEN_OFFSET) as *mut u64;
+            // `vable_token` is `usize` (pointer-width: 4 on wasm32). Writing
+            // 8 bytes would clobber the following field.
+            let token_ptr = ptr.add(PYFRAME_VABLE_TOKEN_OFFSET) as *mut usize;
             *token_ptr = 0;
         }
     }
